[TR-17] Creator can delete input
The creator of a comment, date option and destination can delete their
input. This option is not visible to the non-creator participant.

Creator is the trip participant, so that it is easier to isolate various
content per trip rather than per user. Creator id is added when the
date option and destination is created.